requirements.txt
streamlit
pandas
requests
beautifulsoup4
yagmail
keyring
python-dotenv